src/requirements.txt
psycopg2-binary==2.9.10
asyncpg==0.30.0
mysql-connector-python==9.1.0

